+++ mii/deployment.py
@@ -10,7 +10,8 @@ import mii
 from deepspeed.launcher.runner import fetch_hostfile
 
 from .constants import DeploymentType, MII_MODEL_PATH_DEFAULT, MODEL_PROVIDER_MAP
-from .utils import logger, get_task_name, get_provider_name
+from .logging import logger
+from .utils import get_task_name, get_provider_name
 from .models.score import create_score_file
 from .models import load_models
 from .config import ReplicaConfig, LoadBalancerConfig
@@ -25,6 +26,7 @@ def deploy(task,
            enable_zero=False,
            ds_config=None,
            mii_config={},
+           instance_type="Standard_NC12s_v3",
            version=1):
     """Deploy a task using specified model. For usage examples see:
 
@@ -139,10 +141,7 @@ def deploy(task,
                           lb_config=lb_config)
 
     if deployment_type == DeploymentType.AML:
-        _deploy_aml(deployment_name=deployment_name,
-                    model_name=model,
-                    task_name=task_name,
-                    version=version)
+        _deploy_aml(deployment_name=deployment_name, instance_type=instance_type, version=version)
     elif deployment_type == DeploymentType.LOCAL:
         return _deploy_local(deployment_name, model_path=model_path)
     elif deployment_type == DeploymentType.NON_PERSISTENT:
@@ -165,12 +164,18 @@ def _deploy_local(deployment_name, model_path):
     mii.utils.import_score_file(deployment_name).init()
 
 
-def _deploy_aml(deployment_name, model_name, task_name, version):
+def _deploy_aml(deployment_name, instance_type, version):
     acr_name = mii.aml_related.utils.get_acr_name()
+    configs = mii.utils.import_score_file(deployment_name, DeploymentType.AML).configs
+    model_name = configs.get("model_name")
+    task_name = configs.get("task_name")
+    replica_num = configs.get("mii_configs").get("replica_num")
     mii.aml_related.utils.generate_aml_scripts(acr_name=acr_name,
                                                deployment_name=deployment_name,
                                                model_name=model_name,
                                                task_name=task_name,
+                                               replica_num=replica_num,
+                                               instance_type=instance_type,
                                                version=version)
     print(
         f"AML deployment assets at {mii.aml_related.utils.aml_output_path(deployment_name)}"
