+++ mii/terminate.py
@@ -5,10 +5,11 @@
 import grpc
 
 import mii
+from mii.logging import logger
 
 
 def terminate(deployment_name):
-    mii.utils.logger.info(f"Terminating server for {deployment_name}")
+    logger.info(f"Terminating server for {deployment_name}")
     generator = mii.mii_query_handle(deployment_name)
     if (deployment_name in mii.non_persistent_models):
         generator.terminate()
@@ -17,7 +18,7 @@ def terminate(deployment_name):
         generator.query({'query': ''})
     except grpc.aio._call.AioRpcError as error:
         if error._code == grpc.StatusCode.UNAVAILABLE:
-            mii.utils.logger.warn(f"Server for {deployment_name} not found")
+            logger.warn(f"Server for {deployment_name} not found")
         else:
             pass
     except (KeyError, TypeError) as error:
